requirements.txt
streamlit
langchain
pandas
python-dotenv
openpyxl
langchain-community
langchain-google-genai
matplotlib